Move checkout DTOs to workflows-dto (Orchestration owns checkout)
The checkout-token endpoint and intent resolution are Orchestration
concerns (api/workflows), so their contracts belong in
@shipfox/api-workflows-dto alongside the step protocol DTOs — not in
@shipfox/api-runners-dto, which is the Scheduling contract package and
is checkout-agnostic. The schemas had no consumers yet, so this is a
pure relocation done before ENG-406 wires a route to them.

- Move checkoutIntentSchema (now its own checkout.ts) and the
  checkout-token schemas into workflows-dto; add the package's first
  Vitest setup + the moved tests.
- Strip both from runners-dto; request-job.ts keeps only the live
  jobPayload schemas (still validated by apps/runner), now with their
  own test.
- Changeset moves to @shipfox/api-workflows-dto.

Refs ENG-402
