+++ .changeset/eng-402-checkout-dto.md
@@ -1,5 +1,5 @@
 ---
-"@shipfox/api-runners-dto": minor
+"@shipfox/api-workflows-dto": minor
 ---
 
-Adds the credential-free checkout intent type and a checkout-token response DTO (basic/bearer auth with a validated expiry), establishing the contract for runner-managed repository checkouts over the lease-scoped checkout flow.
+Adds the credential-free checkout intent type and a checkout-token response DTO (basic/bearer auth with a validated expiry) to the Orchestration DTO package, establishing the contract for runner-managed repository checkouts over the lease-scoped checkout flow.
