+++ libs/api/runners-dto/src/index.ts
@@ -1,15 +1,9 @@
 export {
-  type CheckoutIntentDto,
-  type CheckoutTokenAuthDto,
-  type CheckoutTokenResponseDto,
   type ClaimedJobResponseDto,
   type CompleteJobBodyDto,
   type CompleteJobResponseDto,
   type CreateRunnerTokenBodyDto,
   type CreateRunnerTokenResponseDto,
-  checkoutIntentSchema,
-  checkoutTokenAuthSchema,
-  checkoutTokenResponseSchema,
   claimedJobResponseSchema,
   completeJobBodySchema,
   completeJobResponseSchema,
