+++ libs/api/runners-dto/src/schemas/request-job.ts
@@ -10,16 +10,6 @@ export const jobPayloadStepSchema = z.object({
 
 export type JobPayloadStepDto = z.infer<typeof jobPayloadStepSchema>;
 
-export const checkoutIntentSchema = z.object({
-  repository_url: z.string().min(1),
-  ref: z.string().min(1),
-  provider: z.string().min(1),
-  source_connection_id: z.string().uuid(),
-  external_repository_id: z.string().min(1),
-});
-
-export type CheckoutIntentDto = z.infer<typeof checkoutIntentSchema>;
-
 export const jobPayloadSchema = z.object({
   job_id: z.string().uuid(),
   run_id: z.string().uuid(),
