+++ libs/api/workflows-dto/src/index.ts
@@ -1,4 +1,10 @@
 export {
+  type CheckoutIntentDto,
+  type CheckoutTokenAuthDto,
+  type CheckoutTokenResponseDto,
+  checkoutIntentSchema,
+  checkoutTokenAuthSchema,
+  checkoutTokenResponseSchema,
   type JobDto,
   jobDtoSchema,
   type NextStepResponseDto,
