+++ libs/api/workflows-dto/src/schemas/index.ts
@@ -1,3 +1,10 @@
+export {type CheckoutIntentDto, checkoutIntentSchema} from './checkout.js';
+export {
+  type CheckoutTokenAuthDto,
+  type CheckoutTokenResponseDto,
+  checkoutTokenAuthSchema,
+  checkoutTokenResponseSchema,
+} from './checkout-token.js';
 export {type JobDto, jobDtoSchema} from './job.js';
 export {
   type NextStepResponseDto,
